+++ include/primitive.hpp
@@ -74,16 +74,16 @@ class primitive : public renderable
 	void set_color( const rgba& n_color);
 	rgba get_color();
 	
-	void set_red( double x);
+	void set_red( float x);
 	double get_red();
 	
-	void set_green( double x);
+	void set_green( float x);
 	double get_green();
 	
-	void set_blue( double x);
+	void set_blue( float x);
 	double get_blue();
 	
-	void set_opacity( double x);
+	void set_opacity( float x);
 	double get_opacity();
 };
